+++ core/src/main/php/peer/http/HttpConnection.class.php
@@ -126,11 +126,14 @@
     /**
      * Send a HTTP request
      *
-     * @param   peer.http.HttpRequest
+     * @param   peer.http.HttpRequest $request
      * @return  peer.http.HttpResponse response object
      */
-    public function send(HttpRequest $r) {
-      return $this->transport->send($r, $this->_timeout, $this->_ctimeout);
+    public function send(HttpRequest $request) {
+      $this->cat && $this->cat->info('>>> ', $request->getHeaderString());
+      $response= $this->transport->send($request, $this->_timeout, $this->_ctimeout);
+      $this->cat && $this->cat->info('<<< ', $response->getHeaderString());
+      return $response;
     }
 
     /**
@@ -268,7 +271,7 @@
      * @param   util.log.LogCategory cat
      */
     public function setTrace($cat) {
-      $this->cat= $cat;
+      $this->transport->setTrace($cat);
     }
   }
 ?>