+++ core/ChangeLog.md
@@ -40,6 +40,8 @@ XP Framework ChangeLog
 
 ### Features
 
+* Added peer.http.HttpConnection::setTrace() for debugging HTTP requests and
+  response headers - (@thekid)
 * Added getParts() method to MultiPart and MimeMessage in peer.mail - (@thekid)
 * Added peer.mail.MailStore::hasFolder() method and implemented it for both
   mail stores, IMAP and Maildir - (@thekid)